Rearrange integration of ucmd.
Created wrapper (command) to catch callbacks from UART rx
ISR. This effectively decouples the module, using an attach/
detach mechanism inside the UART module.

By running command_init the ucmd and line modules are initialized.
Next UART IRQ is enabled and callback is attached.

All is left to do is to poll the buffer in line module for a
complete message by running a command_update function in the
low priority task.