Add error message when ADC calibration fails.
Take care of fake overtemperature error message.

At startup, the interrupt on the overtemperature gpio
was being enabled before pulling low that same line to
start in safe state with driver turned off. Thus when the line
was being pulled low to take care of driver and disable it,
the interrupt was being misfired. Now the interrupt is enabled
after the line is being pulled low.

Update callback mechanism for ADC ISR.

Add a register callback function into ADC module to more cleanly
register a function (fast control) for callback on ADC
end of conversion interrupt. Pass into callback, as void parameter,
the current ADC conversion values (for current control).

Add "shunt resistor unreadable" logic.

When the PWM is over certain threshold (close to 100%),
the phase is not connected to ground enough to create a
current through the shunt resistor. If this is allowed to
happen, then the opposite phase needs to be read.

Extra logic is added to take the average of both phases
when the duty cycle is below the threshold.